requirements/doc.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/doc.txt requirements/doc.in
#
alabaster==0.7.11         # via sphinx
astroid==2.0.4            # via pylint
atomicwrites==1.2.1       # via pytest
attrs==18.2.0             # via pytest
babel==2.6.0              # via sphinx
backcall==0.1.0           # via ipython
bleach==2.1.4             # via nbconvert
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
dask==0.19.1
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via nbconvert
docutils==0.14            # via sphinx
entrypoints==0.2.3        # via nbconvert
factory-boy==2.11.1
faker==0.9.0
graphviz==0.9
html5lib==1.0.1           # via bleach
idna==2.7                 # via requests
imagesize==1.1.0          # via sphinx
invoke==1.1.1
ipykernel==4.9.0          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==6.5.0            # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.4.1         # via jupyter
isort==4.3.4              # via pylint
jedi==0.12.1              # via ipython
jinja2==2.10              # via nbconvert, notebook, sphinx
jsonschema==2.6.0         # via jupyterlab-launcher, nbformat
jupyter-client==5.2.3     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==5.2.0    # via jupyter
jupyter-core==4.4.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
jupyterlab-launcher==0.13.1  # via jupyterlab
jupyterlab==0.34.7
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
mistune==0.8.3            # via nbconvert
more-itertools==4.3.0     # via pytest
nbconvert==5.4.0          # via jupyter, notebook
nbformat==4.4.0           # via ipywidgets, nbconvert, notebook
notebook==5.6.0           # via jupyter, jupyterlab, jupyterlab-launcher, widgetsnbextension
numpy==1.15.1             # via pandas
packaging==17.1           # via sphinx
pandas-mapper==0.1.2
pandas==0.23.4
pandocfilters==1.4.2      # via nbconvert
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pluggy==0.7.1             # via pytest
prometheus-client==0.3.1  # via notebook
prompt-toolkit==1.0.15    # via ipython, jupyter-console
psycopg2==2.7.5
ptyprocess==0.6.0         # via pexpect, terminado
py==1.6.0                 # via pytest
pygments==2.2.0           # via ipython, jupyter-console, nbconvert, qtconsole, sphinx
pylint==2.1.1
pyparsing==2.2.0          # via packaging
pytest==3.8.0
python-dateutil==2.7.3    # via faker, jupyter-client, pandas
pytz==2018.5              # via babel, pandas
pyzmq==17.1.2             # via jupyter-client, notebook
qtconsole==4.4.1          # via jupyter
requests==2.19.1          # via sphinx
send2trash==1.5.0         # via notebook
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via astroid, bleach, faker, html5lib, more-itertools, packaging, prompt-toolkit, pytest, python-dateutil, sphinx, traitlets
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.4.1
sphinx==1.7.9
sphinxcontrib-websupport==1.1.0  # via sphinx
sqlalchemy==1.2.11
terminado==0.8.1          # via notebook
testpath==0.3.1           # via nbconvert
text-unidecode==1.2       # via faker
tornado==5.1              # via ipykernel, jupyter-client, notebook, terminado
traitlets==4.3.2          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.1.0          # via astroid
urllib3==1.23             # via requests
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
widgetsnbextension==3.4.1  # via ipywidgets
wrapt==1.10.11            # via astroid
